ajoute form validatin page sign up et afficher les commandes

+++ app/controllers/ProduitController.php
@@ -7,6 +7,8 @@ class ProduitController extends Controller{
 
     
     public function addProduit(){
+
+        if($_SESSION['role'] == 'admin'){
         $data = $_POST;
         $target_img = dirname(APPROOT) . "/public/img/image_produit/" . $_FILES['image']['name'];
         move_uploaded_file($_FILES['image']['tmp_name'], $target_img);
@@ -19,24 +21,32 @@ class ProduitController extends Controller{
        
         redirect("/pages/table_produits");
     }
+    else{
+        redirect("/pages/index");
+    }
+
+}
 
     public function affichage_produit(){
+
         
          $data =  $this->produitModel->affichage_produit();
-        
-        
         
     }
     
     public function delete_produit($id){
         
+        if($_SESSION['role'] == 'admin'){
         $this->produitModel->delete_produit($id);
         
         redirect("/pages/table_produits");
+        }else{
+            redirect("/pages/index");
+        }
     } 
 
     public function update_produit($id){
-     
+        if($_SESSION['role'] == 'admin'){
         if($_SERVER['REQUEST_METHOD'] == 'GET'){
             $produit = $this->produitModel->get_produit($id);
             $this->view('pages/update_produit',(array) $produit);
@@ -51,11 +61,14 @@ class ProduitController extends Controller{
             $this->produitModel->update_produit($data,$id);
             redirect('/pages/table_produits');
          }
+        }else{
+            redirect("/pages/index");
+        }
         }
 
-        public function affichage_count(){
-            $this->produitModel->affichage_count();
+        // public function affichage_count(){
+        //     $this->produitModel->affichage_count();
            
-        }
+        // }
 }
 ?>
